# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/students": {
      "get": {
        "description": "Fetch a list of students with pagination. Filter by courseId, or semesterId.",
        "operationId": "UsersController_getStudents",
        "parameters": [
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "description": "Filter by Student ID",
            "schema": {
              "example": "20261144",
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "description": "Filter by Course ID",
            "schema": {
              "example": "50000004",
              "type": "string"
            }
          },
          {
            "name": "semesterId",
            "required": false,
            "in": "query",
            "description": "Filter by Semester ID",
            "schema": {
              "example": "20000008",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by Student Status",
            "schema": {
              "example": "active",
              "type": "string",
              "enum": [
                "active",
                "passive",
                "suspended",
                "graduated"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "The page number for pagination context",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of rows to return per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully"
          },
          "403": {
            "description": "Forbidden access"
          }
        },
        "summary": "Get List of Filtered Students",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Retrieve detailed data for a specific student, basic info from users table and the semester maxCredits.",
        "operationId": "UsersController_findUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "20261144",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": "User was not found"
          }
        },
        "summary": "Get Student Detailed Data",
        "tags": [
          "users"
        ]
      }
    },
    "/users/gpa/{studentId}/{semesterId}": {
      "get": {
        "operationId": "UsersController_getStudentCurrentGPA",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "20261144",
              "type": "string"
            }
          },
          {
            "name": "semesterId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "semester-1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/courses/add": {
      "post": {
        "description": "Creates a new course in the system.",
        "operationId": "CoursesController_createCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "409": {
            "description": "Course with this code already exists"
          }
        },
        "summary": "Create Course",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/courses": {
      "get": {
        "description": "Fetch a list of courses. Filter by code, name, or status.",
        "operationId": "CoursesController_getCourses",
        "parameters": [
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Filter by course code",
            "schema": {
              "example": "CSE101",
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": false,
            "in": "query",
            "description": "Filter by course title",
            "schema": {
              "example": "Introduction to Computer Science",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by Status (active - true/false)",
            "schema": {
              "example": "true",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by course status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Get Filtered Courses",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/offered-courses": {
      "get": {
        "description": "Fetch offered courses for the current semester that the authenticated student can enroll in based on remaining credit limits.",
        "operationId": "OfferedCoursesController_getOfferedCoursesForStudent",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offered courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": "No active semester was found"
          }
        },
        "summary": "Get Available Offered Courses",
        "tags": [
          "OfferedCourses"
        ]
      }
    },
    "/offered-courses/admin": {
      "get": {
        "description": "Fetch all offered courses, optionally filtered by semester, for administrative management.",
        "operationId": "OfferedCoursesController_getOfferedCoursesForAdmin",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "semesterId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "20000008",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offered courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Get Offered Courses for Admin",
        "tags": [
          "OfferedCourses"
        ]
      }
    },
    "/offered-courses/add": {
      "post": {
        "description": "Offer a course in a specific semester.",
        "operationId": "OfferedCoursesController_createOfferedCourse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferedCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course added to semester successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": "Course or semester not found"
          },
          "409": {
            "description": "This course is already offered in the selected semester"
          }
        },
        "summary": "Add Offered Course",
        "tags": [
          "OfferedCourses"
        ]
      }
    },
    "/offered-courses/{id}": {
      "delete": {
        "description": "Remove a course from a semester offering.",
        "operationId": "OfferedCoursesController_deleteOfferedCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offered course removed successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          },
          "404": {
            "description": "Offered course not found"
          }
        },
        "summary": "Delete Offered Course",
        "tags": [
          "OfferedCourses"
        ]
      }
    },
    "/offered-courses/available-courses/{semesterId}": {
      "get": {
        "description": "Get courses that can be added to a semester offering (not already offered)",
        "operationId": "OfferedCoursesController_getAvailableCoursesForSemester",
        "parameters": [
          {
            "name": "semesterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available courses fetched successfully"
          },
          "403": {
            "description": "You are not authorized to access"
          }
        },
        "summary": "Available Courses for Semester",
        "tags": [
          "OfferedCourses"
        ]
      }
    }
  },
  "info": {
    "title": "APIs Documentation for EasyLearn Web App",
    "description": "APIs description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Course Code",
            "example": "PHY101"
          },
          "title": {
            "type": "string",
            "description": "Course Title",
            "example": "Introduction to Physics"
          },
          "credit": {
            "type": "number",
            "description": "Course Credits",
            "example": 3
          },
          "active": {
            "type": "boolean",
            "description": "Status (active)",
            "example": true
          }
        },
        "required": [
          "code",
          "title",
          "credit"
        ]
      },
...
      "CreateOfferedCourseDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "description": "The unique course identifier",
            "example": "50000004"
          },
          "semesterId": {
            "type": "string",
            "description": "The unique semester identifier",
            "example": "20000008"
          }
        },
        "required": [
          "courseId",
          "semesterId"
        ]
      }
    }
  }
}